debug

+++ cg/utils/files.py
@@ -30,7 +30,11 @@ def get_files_matching_pattern(directory: Path, pattern: str) -> list[Path]:
 
 def copy_file(file_path: Path, destination: Path):
     """Copy a file to a destination."""
+    if not file_path.exists():
+        raise FileNotFoundError
     shutil.copy(src=file_path, dst=destination)
+    if not Path(destination, file_path.name).exists():
+        raise FileNotFoundError
 
 
 def get_all_files_in_dir(base_path: Path) -> set[Path]:
